Factor out Consensus' Examples data type into a separate module
This required to factor out 'SomeResult' as well.
