+++ ouroboros-consensus/src/unstable-consensus-testlib/Test/Util/Serialisation/Golden.hs
@@ -30,18 +30,11 @@
 -- In particular, if we introduce golden tests in new suites, we need to add
 -- a line in the nix configuration above similar to the previous ones.
 module Test.Util.Serialisation.Golden (
-    Examples (..)
-  , Labelled
-  , ToGoldenDirectory (..)
-  , combineExamples
+    ToGoldenDirectory (..)
   , goldenTest_SerialiseDisk
   , goldenTest_SerialiseNodeToClient
   , goldenTest_SerialiseNodeToNode
   , goldenTest_all
-  , labelled
-  , mapExamples
-  , prefixExamples
-  , unlabelled
   ) where
 
 import           Cardano.Prelude (forceElemsToWHNF)
@@ -58,16 +51,10 @@ import           Data.List (nub)
 import qualified Data.Map.Strict as Map
 import           Data.Proxy (Proxy (..))
 import           GHC.Stack (HasCallStack)
-import           Ouroboros.Consensus.Block (BlockProtocol, CodecConfig, Header,
-                     HeaderHash, SlotNo, SomeSecond)
-import           Ouroboros.Consensus.HeaderValidation (AnnTip)
-import           Ouroboros.Consensus.Ledger.Abstract (LedgerState)
-import           Ouroboros.Consensus.Ledger.Extended (ExtLedgerState,
-                     encodeExtLedgerState)
-import           Ouroboros.Consensus.Ledger.Query (BlockQuery, QueryVersion,
+import           Ouroboros.Consensus.Block (CodecConfig)
+import           Ouroboros.Consensus.Ledger.Extended (encodeExtLedgerState)
+import           Ouroboros.Consensus.Ledger.Query (QueryVersion,
                      nodeToClientVersionToQueryVersion)
-import           Ouroboros.Consensus.Ledger.SupportsMempool (ApplyTxErr, GenTx,
-                     GenTxId)
 import           Ouroboros.Consensus.Node.NetworkProtocolVersion
                      (HasNetworkProtocolVersion (..),
                      SupportedNetworkProtocolVersion (..))
@@ -77,18 +64,16 @@ import           Ouroboros.Consensus.Node.Run (SerialiseDiskConstraints,
 import           Ouroboros.Consensus.Node.Serialisation
                      (SerialiseNodeToClient (..), SerialiseNodeToNode (..),
                      SerialiseResult (..))
-import           Ouroboros.Consensus.Protocol.Abstract (ChainDepState)
-import           Ouroboros.Consensus.Storage.Serialisation (EncodeDisk (..),
-                     SerialisedHeader)
+import           Ouroboros.Consensus.Storage.Serialisation (EncodeDisk (..))
 import           Ouroboros.Consensus.Util.CBOR (decodeAsFlatTerm)
 import           Ouroboros.Consensus.Util.Condense (Condense (..))
-import           Ouroboros.Network.Block (Serialised)
 import           System.Directory (createDirectoryIfMissing)
 import           System.FilePath (takeDirectory, (</>))
 import           Test.Cardano.Ledger.Binary.TreeDiff (CBORBytes (..), diffExpr)
 import           Test.Tasty
 import           Test.Tasty.Golden.Advanced (goldenTest)
-import           Test.Util.Serialisation.Roundtrip (SomeResult (..))
+import           Test.Util.Serialisation.Examples (Examples (..), Labelled)
+import           Test.Util.Serialisation.SomeResult (SomeResult (..))
 
 {-------------------------------------------------------------------------------
   Golden test
@@ -203,112 +188,6 @@ goldenTests testName examples enc goldenFolder
   where
     labels :: [Maybe String]
     labels = map fst examples
-
-{-------------------------------------------------------------------------------
-  Examples
--------------------------------------------------------------------------------}
-
-type Labelled a = [(Maybe String, a)]
-
-unlabelled :: a -> Labelled a
-unlabelled x = [(Nothing, x)]
-
-labelled :: [(String, a)] -> Labelled a
-labelled = map (first Just)
-
-data Examples blk = Examples {
-      exampleBlock            :: Labelled blk
-    , exampleSerialisedBlock  :: Labelled (Serialised blk)
-    , exampleHeader           :: Labelled (Header blk)
-    , exampleSerialisedHeader :: Labelled (SerialisedHeader blk)
-    , exampleHeaderHash       :: Labelled (HeaderHash blk)
-    , exampleGenTx            :: Labelled (GenTx blk)
-    , exampleGenTxId          :: Labelled (GenTxId blk)
-    , exampleApplyTxErr       :: Labelled (ApplyTxErr blk)
-    , exampleQuery            :: Labelled (SomeSecond BlockQuery blk)
-    , exampleResult           :: Labelled (SomeResult blk)
-    , exampleAnnTip           :: Labelled (AnnTip blk)
-    , exampleLedgerState      :: Labelled (LedgerState blk)
-    , exampleChainDepState    :: Labelled (ChainDepState (BlockProtocol blk))
-    , exampleExtLedgerState   :: Labelled (ExtLedgerState blk)
-    , exampleSlotNo           :: Labelled SlotNo
-    }
-
-emptyExamples :: Examples blk
-emptyExamples = Examples {
-      exampleBlock            = mempty
-    , exampleSerialisedBlock  = mempty
-    , exampleHeader           = mempty
-    , exampleSerialisedHeader = mempty
-    , exampleHeaderHash       = mempty
-    , exampleGenTx            = mempty
-    , exampleGenTxId          = mempty
-    , exampleApplyTxErr       = mempty
-    , exampleQuery            = mempty
-    , exampleResult           = mempty
-    , exampleAnnTip           = mempty
-    , exampleLedgerState      = mempty
-    , exampleChainDepState    = mempty
-    , exampleExtLedgerState   = mempty
-    , exampleSlotNo           = mempty
-    }
-
-combineExamples ::
-     forall blk.
-     (forall a. Labelled a -> Labelled a -> Labelled a)
-  -> Examples blk
-  -> Examples blk
-  -> Examples blk
-combineExamples f e1 e2 = Examples {
-      exampleBlock            = combine exampleBlock
-    , exampleSerialisedBlock  = combine exampleSerialisedBlock
-    , exampleHeader           = combine exampleHeader
-    , exampleSerialisedHeader = combine exampleSerialisedHeader
-    , exampleHeaderHash       = combine exampleHeaderHash
-    , exampleGenTx            = combine exampleGenTx
-    , exampleGenTxId          = combine exampleGenTxId
-    , exampleApplyTxErr       = combine exampleApplyTxErr
-    , exampleQuery            = combine exampleQuery
-    , exampleResult           = combine exampleResult
-    , exampleAnnTip           = combine exampleAnnTip
-    , exampleLedgerState      = combine exampleLedgerState
-    , exampleChainDepState    = combine exampleChainDepState
-    , exampleExtLedgerState   = combine exampleExtLedgerState
-    , exampleSlotNo           = combine exampleSlotNo
-    }
-  where
-    combine :: (Examples blk -> Labelled a) -> Labelled a
-    combine getField = f (getField e1) (getField e2)
-
-instance Semigroup (Examples blk) where
-  (<>) = combineExamples (<>)
-
-instance Monoid (Examples blk) where
-  mempty  = emptyExamples
-  mappend = (<>)
-
-mapExamples ::
-     forall blk.
-     (forall a. Labelled a -> Labelled a)
-  -> Examples blk
-  -> Examples blk
-mapExamples f = combineExamples (const f) mempty
-
--- | Add the given prefix to each labelled example.
---
--- When a label is empty, the prefix is used as the label. If the label is not
--- empty, the prefix and @_@ are prepended.
-prefixExamples :: String -> Examples blk -> Examples blk
-prefixExamples prefix = mapExamples addPrefix
-  where
-    addPrefix :: Labelled a -> Labelled a
-    addPrefix l = [
-          (Just label, x)
-        | (mbLabel, x) <- l
-        , let label = case mbLabel of
-                Nothing  -> prefix
-                Just lbl -> prefix <> "_" <> lbl
-        ]
 
 {-------------------------------------------------------------------------------
   Skeletons
